refactoring tests & improving on-chain query coverage
