+++ tests/uniswap_v3/test_pool_swaps.py
@@ -1,3 +1,4 @@
+# type: ignore
 import json
 import logging
 import math
